Move addNode to a different file
